Build tests JAR, rename test data package
The contained test data might be useful to use for other unit and integration
tests in the ConfAPI plugins. For that reason, a test JAR containing tests and
test data is now created and deployed together with the library itself.

Furthermore, the test data package for Atlassian mail tests has been renamed as
it has been missing the com prefix.