Implement unit test assert class for REST resources
The assertions for testing a REST resource class are quite complicated, require
reflections and result in a too large number of code lines. This is the reason
why the class ResourceAssert has been implemented that allows checking for paths
and HTTP methods to be correct in a faster way than integration tests.

Currently, the assertions are only available in the test-jar.