Define bean naming conventions, add tests, fixes
This commit is the first step in defining the naming convention for the ConfAPI
artifacts. It defines how the model beans have to be named and provides
automated tests for checking these naming conventions just by extending from the
AbstractBeanTest class in the bean's test class.

This commit also includes some smaller fixes in comments or regarding
standardization.