Move resource implementation to ConfAPI Commons
During the process of standardizing the ConfAPI plugins, more and more of the
REST and the underlying service interfaces have moved to the Commons library.
This lead to the current state that the actual (default) implementation of the
resources also can be moved to the Commons library, while the actual service
implementations should be moved (back) to the actual plugins in order to get rid
of fixed Atlassian dependencies in the Commons library (see #13).

This commit contains two PoC's for REST resources that are now implemented in
the Commons library. First the PingResource that does not contain any services
at all and second the DirectoriesResource that uses a custom service only.

While the DirectoryService interface will stay in the Commons library, the
DirectoryServiceImpl will be moved to the actual plugins after.

In order to use one of the here implemented in an actual plugin, the developer
only needs to extend from the Abstract<Entity>ResourceImpl, inject the
service(s) and add the corresponding `@Path` to the class.

Snapshot version bump because of interface change in DirectoriesResource.