Remove Atlassian deps from service interfaces
Replace Atlassian exceptions in service interfaces by new custom exceptions in
order to get rid of the Atlassian dependencies in the ConfAPI Commons library.

Remove Javadoc where it is not necessary which is the case for resource
interfaces that are documented by Swagger annotations and service implementation
classes where the Javadoc belongs to the service interfaces.

The ValidationException which is a RuntimeException of the BeanValidationUtil
has been replaced by a BadRequestException.

Adjust tests for new exception behaviour.

Snapshot version bump due to API changes.