License resource with addLicense method
Due to a resource addressing conflict (path + mediatype) the
license resource was redesigned. Now, there is a dedicated
POST addLicense resource which adds a new license to an already
existing license collection.

The PUT setLicenses resource replaces the existing license
collection.

Because of this redesign the 'clear' flag was removed. Clearing
license details is achieved with the setLicenses method.