Rename packages, rename group ID, version bump, closes #25
As this library does not contain any Atlassian dependenies anymore and as the
Atlassian brand guidelines also require so, the packages have been renamed so
they don't contain 'atlassian' anymore, but know end with 'commons' to make more
clear where the classes are coming from.

The same is for the group ID, which ended with '.atlassian' which has been
removed again.

Snapshot version bump due to build breaking changes.