Abstract licenses resource integration tests implementation
This commit introduces abstract integration testing for the licenses
resource. In addition, a sample LicenseBean with a generic 3h developer
license is provided as well. The key might need to be replaced on
a regular basis.

Also, the abstract classes for 'ping' and 'settings' resource integration
tests were renamed. For ping, the mediatype was corrected to text/plain.