Use object data types instead of primitive data types
Primitive data types in bean classes should ONLY be used, if a default value is
specified or assumed. In all other cases, having primitive data types can lead
to a misconfiguration as e.g. a `boolean` property that is NOT given in the JSON
data will always be evaluated to `false` while a `Boolean` property will
evaluate to `null` and either should be replaced by the wanted default value or
just not be configured at all.

Same for `int` and `long`.