Bean Validation Extension
Upon usage of the method `BeanValidationUtil.validate` in Bitbucket it was recognized that the required
javax.el dependencies cannot be provided as maven dependencies for runtime. However, a bitbucket internal
validator must be injected. This commit extracts the validator validation set processing for other
plugins to use when the default validator of the commons lib cannot be used.