Improve WebApplicationException mapping
As the WebApplicationException needs to be instantiated by a Throwable and there
is no way to instantiate it with a String message only, the call `getMessage()`
always returns the `cause.toString` which always contains the exception class
name as a prefix for the actual message.

With this change, this exception class name is always removed from the exception
message again so that ErrorCollections don't contain any technical details
anymore.