Upgrade Crowd to version 5.x
This also requires to set some fixed version and to pull in some new
dependencies, due to the use of the dependency management and  because
the transitive version updates of dependencies would require a large
set of changes.

Also, the spring scanner configuration needed to be fixed, otherwise
Crowd 5 wouldn't load the plugin.