Allow setting users together with directories, improve users endpoint
Users (and groups) directly belong to a directory. Because of that, the
cleanest way to declaratively add or update directory users (and groups)
would be to set them as a child element of a directory. E.g.:

```
directories:
  internal:
    name: Internal Directory
    ...
    users:
      user1:
        password: ...
      user2:
        password: ...
```

With this change it is possible to set users together with directories,
however still as lists (not as maps as in the example), and for groups
this still needs to be implemented if needed.

Also improve the users endpoint itself and allow passing a directory ID
when performing user actions. The old behaviour that didn't require a
directory ID is inacurate because would just any random directory that
contains a user with the given name, and thus the related methods have
been deprecated.