Fix groups endpoint collision by changing endpoint method
The plugin didn't work anymore because the groups endpoint had two
sub-endpoints that produced the same media type (used the same method).
In order to avoid the collision, the `PATCH` method has been introduced
in ConfAPI commons (as it's missing in the old JAX-RS library) that is
now used for the `setGroups` endpoint.

The usage of this method for `setGroups` is now inconsistent with the
existing `set<OBJECT>s` endpoints, however we will need to streamline
the existing endpoints anyway and will change over existing endpoints
to `PATCH` in a breaking major release also.

In general, `PATCH` is supposed to work a bit different than how we use
it now. However, our already existing update endpoints (`PUT`) already
supports updating only modified values such as `PATCH` would do. On the
other hand, we could consider the whole instance, e.g. Crowd to be
patched, so the method `PATCH` might be a reasonable choice.