Remove backup / restore endpoint
In the past, the backup / restore endpoint has been implemented in
ConfAPI to be able to perform export / import operations via REST API,
which was long a missing feature in Confluence. The endpoint was fully
working, but never really fully finished.

In the last years, Atlassian has finally implemented an own official
backup / restore endpoint. In order to achieve this, they also fully
reimplemented the whole related internal Java API, and deprecated the
whole related old Java API.

Porting the existing backend / restore endpoint onto the new Java API
would be equal to a full reimplementation. For this reason, the endpoint
that has always been 'experimental' is removed again in favor of the
official endpoint.