Make REST API method names more consistent
In entity collections such as application links, where each entity
itself is a complex construct already, we use the following naming
scheme together with the following HTTP methods.

Single entity:

- `GET`:    getEntity
- `POST`:   createEntity
- `PUT`:    updateEntity
- `DELETE`: deleteEntity

Multiple entities:

- `GET`:    getEntities
- `PUT`:    setEntities
- `DELETE`: deleteEntities

Whereas `PUT` setEntities will be replaced by `PATCH` in a future
release to better explain that these methods may create OR update
entities, but only the ones that are send in the request.

For collections that kind of form a single entity themselves, the
methods work a bit different. E.g. the trusted proxies in Crowd are one
single entity in the form of a list of entries (the trusted proxies).
Here, and in similar cases, we would use:

- `GET`:    getEntries
- `PUT`:    setEntries
- `POST`:   addEntry
- `DELETE`: removeEntry

However, here we would stick to `PUT` for setEntries since it makes more
sense.